refactor: consolidate types under src
