+++ __tests__/components/Navbar.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import Navbar from '@/components/Navbar';
-import { Category } from '../../types';
+import { Category } from '@/types';
 import { useRouter } from 'next/router';
 import '@testing-library/jest-dom';
