+++ src/components/FeaturedProductsCarousel.tsx
@@ -1,6 +1,6 @@
 // components/FeaturedProductsCarousel.tsx
 import Image from 'next/image';
-import type { Product } from '../../types'; // Import the shared Product type
+import type { Product } from '@/types'; // Import the shared Product type
 import styles from '../../styles/FeaturedProductsCarousel.module.css'; // Import CSS module
 
 type Props = {
