--- types/order.ts
@@ -1,11 +1,0 @@
-export interface CartItem {
-  productId: number;
-  quantity: number;
-}
-
-export interface CheckoutSession {
-  id: string;
-  url?: string;
-  customer: { name?: string | null } | null;
-  customer_details?: { email?: string | null; name?: string | null } | null;
-}
